fix(server): address PR review feedback
- board.rs: fix regex pattern to match DB constraint {0,63}
- pagination.rs: fix overflow by casting to u64 before multiply
- invoker.rs: remove unused InvokeError::NotFound variant
- messages.rs: fix total count extraction in search_by_markers
- scratchpad.rs: fix spawn_cleanup lifetime with explicit PgPool clone
- extractors.rs: fix misleading error mapping to InvalidFormat
- health.rs: fix test to unwrap Json wrapper
- marker.rs: fix case-insensitivity for float. prefix parsing

+++ floatctl-server/src/models/pagination.rs
@@ -30,8 +30,9 @@ impl Pagination {
     }
 
     /// Calculate SQL OFFSET value.
+    /// Cast to u64 before multiplication to prevent overflow.
     pub fn offset(&self) -> u64 {
-        ((self.page - 1) * self.per_page) as u64
+        (self.page as u64 - 1) * self.per_page as u64
     }
 
     /// Get LIMIT value.
